+++ hefquin-engine/src/main/java/se/liu/ida/hefquin/engine/queryplan/physical/PhysicalPlan.java
@@ -3,6 +3,7 @@ package se.liu.ida.hefquin.engine.queryplan.physical;
 import java.util.NoSuchElementException;
 
 import se.liu.ida.hefquin.base.query.ExpectedVariables;
+import se.liu.ida.hefquin.engine.queryplan.info.QueryPlanningInfo;
 
 public interface PhysicalPlan
 {
@@ -32,4 +33,11 @@ public interface PhysicalPlan
 	 * then a {@link NoSuchElementException} will be thrown.
 	 */
 	PhysicalPlan getSubPlan( int i ) throws NoSuchElementException;
+
+	/**
+	 * Returns an object that captures query-planning-related
+	 * information about this plan. This object is meant to be
+	 * requested and populated by the query planner. 
+	 */
+	QueryPlanningInfo getQueryPlanningInfo();
 }