+++ hefquin-engine/src/main/java/se/liu/ida/hefquin/engine/queryplan/physical/impl/PhysicalPlanWithUnaryRootImpl.java
@@ -3,12 +3,14 @@ package se.liu.ida.hefquin.engine.queryplan.physical.impl;
 import java.util.NoSuchElementException;
 
 import se.liu.ida.hefquin.base.query.ExpectedVariables;
+import se.liu.ida.hefquin.engine.queryplan.info.QueryPlanningInfo;
 import se.liu.ida.hefquin.engine.queryplan.physical.PhysicalPlan;
 import se.liu.ida.hefquin.engine.queryplan.physical.PhysicalPlanWithUnaryRoot;
 import se.liu.ida.hefquin.engine.queryplan.physical.UnaryPhysicalOp;
 import se.liu.ida.hefquin.engine.queryplan.utils.PhysicalPlanFactory;
 
-public class PhysicalPlanWithUnaryRootImpl implements PhysicalPlanWithUnaryRoot
+public class PhysicalPlanWithUnaryRootImpl extends BaseForPhysicalPlan
+                                           implements PhysicalPlanWithUnaryRoot
 {
 	private final UnaryPhysicalOp rootOp;
 	private final PhysicalPlan subPlan;
@@ -19,6 +21,28 @@ public class PhysicalPlanWithUnaryRootImpl implements PhysicalPlanWithUnaryRoot
 	 */
 	protected PhysicalPlanWithUnaryRootImpl( final UnaryPhysicalOp rootOp,
 	                                         final PhysicalPlan subPlan ) {
+		assert rootOp != null;
+		assert subPlan != null;
+
+		this.rootOp = rootOp;
+		this.subPlan = subPlan;
+	}
+
+	/**
+	 * Instead of creating such a plan directly using
+	 * this constructor, use {@link PhysicalPlanFactory}.
+	 * <p>
+	 * This constructor should be used only if the plan is meant to be
+	 * constructed with an already existing {@link QueryPlanningInfo}
+	 * object. Since this object may later be extended with additional
+	 * properties for this plan, it is important not to create multiple
+	 * plans with the same {@link QueryPlanningInfo} object.
+	 */
+	protected PhysicalPlanWithUnaryRootImpl( final UnaryPhysicalOp rootOp,
+	                                         final QueryPlanningInfo qpInfo,
+	                                         final PhysicalPlan subPlan ) {
+		super(qpInfo);
+
 		assert rootOp != null;
 		assert subPlan != null;