+++ hefquin-engine/src/main/java/se/liu/ida/hefquin/engine/queryplan/utils/PhysicalPlanFactory.java
@@ -5,6 +5,7 @@ import java.util.List;
 
 import se.liu.ida.hefquin.base.data.VocabularyMapping;
 import se.liu.ida.hefquin.base.query.TriplePattern;
+import se.liu.ida.hefquin.engine.queryplan.info.QueryPlanningInfo;
 import se.liu.ida.hefquin.engine.queryplan.logical.BinaryLogicalOp;
 import se.liu.ida.hefquin.engine.queryplan.logical.LogicalOperator;
 import se.liu.ida.hefquin.engine.queryplan.logical.NaryLogicalOp;
@@ -38,10 +39,31 @@ public class PhysicalPlanFactory
 	 */
 	public static PhysicalPlan createPlan( final LogicalOperator lop,
 	                                       final PhysicalPlan... subplans ) {
-		if (      lop instanceof NullaryLogicalOp ) return createPlan( (NullaryLogicalOp) lop );
-		else if ( lop instanceof UnaryLogicalOp )   return createPlan( (UnaryLogicalOp) lop, subplans[0] );
-		else if ( lop instanceof BinaryLogicalOp )  return createPlan( (BinaryLogicalOp) lop, subplans[0], subplans[1] );
-		else if ( lop instanceof NaryLogicalOp )    return createPlan( (NaryLogicalOp) lop, subplans );
+		return createPlan( lop, null, subplans );
+	}
+
+	/**
+	 * Creates a physical plan in which the root operator is the
+	 * default physical operator for the given logical operator,
+	 * as per {@link LogicalToPhysicalOpConverter}. The given
+	 * subplans become children of the root operator.
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the plan is indeed meant to be associated with it (for instance,
+	 * when creating a physical plan for a logical plan that is associated
+	 * with this object). Also, do not create different physical plans with
+	 * the same {@link QueryPlanningInfo} object because these objects may
+	 * later be extended with additional properties for each plan; instead,
+	 * make copies of such an object if needed.
+	 */
+	public static PhysicalPlan createPlan( final LogicalOperator lop,
+	                                       final QueryPlanningInfo qpInfo,
+	                                       final PhysicalPlan... subplans ) {
+		if (      lop instanceof NullaryLogicalOp nullaryLOP ) return createPlan(nullaryLOP, qpInfo);
+		else if ( lop instanceof UnaryLogicalOp unaryLOP )     return createPlan(unaryLOP, qpInfo, subplans[0]);
+		else if ( lop instanceof BinaryLogicalOp binaryLOP )   return createPlan(binaryLOP, qpInfo, subplans[0], subplans[1]);
+		else if ( lop instanceof NaryLogicalOp naryLOP )       return createPlan(naryLOP, qpInfo, subplans);
 		else throw new UnsupportedOperationException("Unsupported type of logical operator: " + lop.getClass().getName() + ".");
 	}
 
@@ -51,13 +73,34 @@ public class PhysicalPlanFactory
 	 * as per {@link LogicalToPhysicalOpConverter}. The given
 	 * subplans become children of the root operator.
 	 */
-	public static PhysicalPlan createPlan( final PhysicalOperator lop,
+	public static PhysicalPlan createPlan( final PhysicalOperator pop,
 	                                       final PhysicalPlan... subplans ) {
-		if (      lop instanceof NullaryPhysicalOp ) return createPlan( (NullaryPhysicalOp) lop );
-		else if ( lop instanceof UnaryPhysicalOp )   return createPlan( (UnaryPhysicalOp) lop, subplans[0] );
-		else if ( lop instanceof BinaryPhysicalOp )  return createPlan( (BinaryPhysicalOp) lop, subplans[0], subplans[1] );
-		else if ( lop instanceof NaryPhysicalOp )    return createPlan( (NaryPhysicalOp) lop, subplans );
-		else throw new UnsupportedOperationException("Unsupported type of physical operator: " + lop.getClass().getName() + ".");
+		return createPlan(pop, null, subplans);
+	}
+
+	/**
+	 * Creates a physical plan in which the root operator is the
+	 * default physical operator for the given logical operator,
+	 * as per {@link LogicalToPhysicalOpConverter}. The given
+	 * subplans become children of the root operator.
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the plan is indeed meant to be associated with it (for instance,
+	 * when creating a physical plan for a logical plan that is associated
+	 * with this object). Also, do not create different physical plans with
+	 * the same {@link QueryPlanningInfo} object because these objects may
+	 * later be extended with additional properties for each plan; instead,
+	 * make copies of such an object if needed.
+	 */
+	public static PhysicalPlan createPlan( final PhysicalOperator pop,
+	                                       final QueryPlanningInfo qpInfo,
+	                                       final PhysicalPlan... subplans ) {
+		if (      pop instanceof NullaryPhysicalOp nullaryPOP ) return createPlan(nullaryPOP, qpInfo);
+		else if ( pop instanceof UnaryPhysicalOp unaryPOP )     return createPlan(unaryPOP, qpInfo, subplans[0]);
+		else if ( pop instanceof BinaryPhysicalOp binaryPOP )   return createPlan(binaryPOP, qpInfo, subplans[0], subplans[1]);
+		else if ( pop instanceof NaryPhysicalOp naryPOP )       return createPlan(naryPOP, qpInfo, subplans);
+		else throw new UnsupportedOperationException("Unsupported type of physical operator: " + pop.getClass().getName() + ".");
 	}
 
 
@@ -84,17 +127,40 @@ public class PhysicalPlanFactory
 	 * Creates a physical plan in which the root operator is the
 	 * default physical operator for the given logical operator,
 	 * as per {@link LogicalToPhysicalOpConverter}.
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the plan is indeed meant to be associated with it (for instance,
+	 * when creating a physical plan for a logical plan that is associated
+	 * with this object). Also, do not create different physical plans with
+	 * the same {@link QueryPlanningInfo} object because these objects may
+	 * later be extended with additional properties for each plan; instead,
+	 * make copies of such an object if needed.
 	 */
-	public static PhysicalPlan createPlan( final NullaryLogicalOp rootOp ) {
+	public static PhysicalPlan createPlan( final NullaryLogicalOp rootOp,
+	                                       final QueryPlanningInfo qpInfo ) {
 		final NullaryPhysicalOp pop = LogicalToPhysicalOpConverter.convert(rootOp);
-		return createPlan(pop);
+		return createPlan(pop, qpInfo);
 	}
 
 	/**
 	 * Creates a physical plan with the given root operator.
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the plan is indeed meant to be associated with it (for instance,
+	 * when creating a physical plan for a logical plan that is associated
+	 * with this object). Also, do not create different physical plans with
+	 * the same {@link QueryPlanningInfo} object because these objects may
+	 * later be extended with additional properties for each plan; instead,
+	 * make copies of such an object if needed.
 	 */
-	public static PhysicalPlan createPlan( final NullaryPhysicalOp rootOp ) {
-		return new PhysicalPlanWithNullaryRootImpl(rootOp) {};
+	public static PhysicalPlan createPlan( final NullaryPhysicalOp rootOp,
+	                                       final QueryPlanningInfo qpInfo ) {
+		if ( qpInfo != null )
+			return new PhysicalPlanWithNullaryRootImpl(rootOp, qpInfo) {};
+		else
+			return new PhysicalPlanWithNullaryRootImpl(rootOp) {};
 	}
 
 
@@ -159,20 +225,43 @@ public class PhysicalPlanFactory
 	 * default physical operator for the given logical operator,
 	 * as per {@link LogicalToPhysicalOpConverter}. The given
 	 * subplan becomes the single child of the root operator.
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the plan is indeed meant to be associated with it (for instance,
+	 * when creating a physical plan for a logical plan that is associated
+	 * with this object). Also, do not create different physical plans with
+	 * the same {@link QueryPlanningInfo} object because these objects may
+	 * later be extended with additional properties for each plan; instead,
+	 * make copies of such an object if needed.
 	 */
 	public static PhysicalPlan createPlan( final UnaryLogicalOp rootOp,
+	                                       final QueryPlanningInfo qpInfo,
 	                                       final PhysicalPlan subplan ) {
 		final UnaryPhysicalOp pop = LogicalToPhysicalOpConverter.convert(rootOp);
-		return createPlan(pop, subplan);
+		return createPlan(pop, qpInfo, subplan);
 	}
 
 	/**
 	 * Creates a physical plan with the given root operator. The
 	 * given subplan becomes the single child of the root operator.
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the plan is indeed meant to be associated with it (for instance,
+	 * when creating a physical plan for a logical plan that is associated
+	 * with this object). Also, do not create different physical plans with
+	 * the same {@link QueryPlanningInfo} object because these objects may
+	 * later be extended with additional properties for each plan; instead,
+	 * make copies of such an object if needed.
 	 */
 	public static PhysicalPlan createPlan( final UnaryPhysicalOp rootOp,
+	                                       final QueryPlanningInfo qpInfo,
 	                                       final PhysicalPlan subplan ) {
-		return new PhysicalPlanWithUnaryRootImpl(rootOp, subplan) {};
+		if ( qpInfo != null )
+			return new PhysicalPlanWithUnaryRootImpl(rootOp, qpInfo, subplan) {};
+		else
+			return new PhysicalPlanWithUnaryRootImpl(rootOp, subplan) {};
 	}
 
 	// --------- plans with binary root operators -----------
@@ -250,12 +339,33 @@ public class PhysicalPlanFactory
 	}
 
 	/**
-	 * Creates a plan with a binary join as root operator.
-	 * The root operator is the default physical operator
+	 * Creates a plan with a binary join as root operator, using
+	 * the default physical operator for such a join (as per
+	 * {@link LogicalToPhysicalOpConverter}).
 	 */
 	public static PhysicalPlan createPlanWithJoin( final PhysicalPlan subplan1,
-													final PhysicalPlan subplan2 ) {
-		return createPlan( LogicalOpJoin.getInstance(), subplan1, subplan2 );
+	                                               final PhysicalPlan subplan2 ) {
+		return createPlanWithJoin(subplan1, subplan2, null);
+	}
+
+	/**
+	 * Creates a plan with a binary join as root operator, using
+	 * the default physical operator for such a join (as per
+	 * {@link LogicalToPhysicalOpConverter}).
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the resulting plan is indeed meant to be associated with it (for
+	 * instance, when creating a physical plan for a logical plan that is
+	 * associated with this object). Also, do not create different physical
+	 * plans with the same {@link QueryPlanningInfo} object because these
+	 * objects may later be extended with additional properties for each plan;
+	 * instead, make copies of such an object if needed.
+	 */
+	public static PhysicalPlan createPlanWithJoin( final PhysicalPlan subplan1,
+	                                               final PhysicalPlan subplan2,
+	                                               final QueryPlanningInfo qpInfo ) {
+		return createPlan( LogicalOpJoin.getInstance(), qpInfo, subplan1, subplan2 );
 	}
 
 	/**
@@ -263,22 +373,45 @@ public class PhysicalPlanFactory
 	 * default physical operator for the given logical operator,
 	 * as per {@link LogicalToPhysicalOpConverter}. The given
 	 * subplans become children of the root operator.
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the plan is indeed meant to be associated with it (for instance,
+	 * when creating a physical plan for a logical plan that is associated
+	 * with this object). Also, do not create different physical plans with
+	 * the same {@link QueryPlanningInfo} object because these objects may
+	 * later be extended with additional properties for each plan; instead,
+	 * make copies of such an object if needed.
 	 */
 	public static PhysicalPlan createPlan( final BinaryLogicalOp rootOp,
+	                                       final QueryPlanningInfo qpInfo,
 	                                       final PhysicalPlan subplan1,
 	                                       final PhysicalPlan subplan2 ) {
 		final BinaryPhysicalOp pop = LogicalToPhysicalOpConverter.convert(rootOp);
-		return createPlan(pop, subplan1, subplan2);
+		return createPlan(pop, qpInfo, subplan1, subplan2);
 	}
 
 	/**
 	 * Creates a physical plan with the given root operator. The
 	 * given subplans become children of the root operator.
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the plan is indeed meant to be associated with it (for instance,
+	 * when creating a physical plan for a logical plan that is associated
+	 * with this object). Also, do not create different physical plans with
+	 * the same {@link QueryPlanningInfo} object because these objects may
+	 * later be extended with additional properties for each plan; instead,
+	 * make copies of such an object if needed.
 	 */
 	public static PhysicalPlan createPlan( final BinaryPhysicalOp rootOp,
+	                                       final QueryPlanningInfo qpInfo,
 	                                       final PhysicalPlan subplan1,
 	                                       final PhysicalPlan subplan2 ) {
-		return new PhysicalPlanWithBinaryRootImpl(rootOp, subplan1, subplan2) {};
+		if ( qpInfo != null )
+			return new PhysicalPlanWithBinaryRootImpl(rootOp, qpInfo, subplan1, subplan2) {};
+		else
+			return new PhysicalPlanWithBinaryRootImpl(rootOp, subplan1, subplan2) {};
 	}
 
 
@@ -289,11 +422,21 @@ public class PhysicalPlanFactory
 	 * default physical operator for the given logical operator,
 	 * as per {@link LogicalToPhysicalOpConverter}. The given
 	 * subplans become children of the root operator.
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the plan is indeed meant to be associated with it (for instance,
+	 * when creating a physical plan for a logical plan that is associated
+	 * with this object). Also, do not create different physical plans with
+	 * the same {@link QueryPlanningInfo} object because these objects may
+	 * later be extended with additional properties for each plan; instead,
+	 * make copies of such an object if needed.
 	 */
 	public static PhysicalPlan createPlan( final NaryLogicalOp rootOp,
+	                                       final QueryPlanningInfo qpInfo,
 	                                       final PhysicalPlan... subplans ) {
 		final NaryPhysicalOp pop = LogicalToPhysicalOpConverter.convert(rootOp);
-		return createPlan(pop, subplans);
+		return createPlan(pop, qpInfo, subplans);
 	}
 
 	/**
@@ -301,29 +444,65 @@ public class PhysicalPlanFactory
 	 * default physical operator for the given logical operator,
 	 * as per {@link LogicalToPhysicalOpConverter}. The given
 	 * subplans become children of the root operator.
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the plan is indeed meant to be associated with it (for instance,
+	 * when creating a physical plan for a logical plan that is associated
+	 * with this object). Also, do not create different physical plans with
+	 * the same {@link QueryPlanningInfo} object because these objects may
+	 * later be extended with additional properties for each plan; instead,
+	 * make copies of such an object if needed.
 	 */
 	public static PhysicalPlan createPlan( final NaryLogicalOp rootOp,
+	                                       final QueryPlanningInfo qpInfo,
 	                                       final List<PhysicalPlan> subplans ) {
 		final NaryPhysicalOp pop = LogicalToPhysicalOpConverter.convert(rootOp);
-		return createPlan(pop, subplans);
+		return createPlan(pop, qpInfo, subplans);
 	}
 
 	/**
 	 * Creates a physical plan with the given root operator. The
 	 * given subplans become children of the root operator.
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the plan is indeed meant to be associated with it (for instance,
+	 * when creating a physical plan for a logical plan that is associated
+	 * with this object). Also, do not create different physical plans with
+	 * the same {@link QueryPlanningInfo} object because these objects may
+	 * later be extended with additional properties for each plan; instead,
+	 * make copies of such an object if needed.
 	 */
 	public static PhysicalPlan createPlan( final NaryPhysicalOp rootOp,
+	                                       final QueryPlanningInfo qpInfo,
 	                                       final PhysicalPlan... subplans ) {
-		return new PhysicalPlanWithNaryRootImpl(rootOp, subplans) {};
+		if ( qpInfo != null )
+			return new PhysicalPlanWithNaryRootImpl(rootOp, qpInfo, subplans) {};
+		else
+			return new PhysicalPlanWithNaryRootImpl(rootOp, subplans) {};
 	}
 
 	/**
 	 * Creates a physical plan with the given root operator. The
 	 * given subplans become children of the root operator.
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the plan is indeed meant to be associated with it (for instance,
+	 * when creating a physical plan for a logical plan that is associated
+	 * with this object). Also, do not create different physical plans with
+	 * the same {@link QueryPlanningInfo} object because these objects may
+	 * later be extended with additional properties for each plan; instead,
+	 * make copies of such an object if needed.
 	 */
 	public static PhysicalPlan createPlan( final NaryPhysicalOp rootOp,
+	                                       final QueryPlanningInfo qpInfo,
 	                                       final List<PhysicalPlan> subplans ) {
-		return new PhysicalPlanWithNaryRootImpl(rootOp, subplans) {};
+		if ( qpInfo != null )
+			return new PhysicalPlanWithNaryRootImpl(rootOp, qpInfo, subplans) {};
+		else
+			return new PhysicalPlanWithNaryRootImpl(rootOp, subplans) {};
 	}
 
 
@@ -411,110 +590,148 @@ public class PhysicalPlanFactory
 	}
 
 	/**
-	 * This function take a inputPlan and unionPlan as input,
-	 * where the unionPlan is required to be a plan with union as root operator, and all subPlans under the UNION are all requests or filters with request.
+	 * This function takes two physical plans as input, with the assumptions
+	 * that the second of these plans i) has a union as its root operator and
+	 * ii) every sub plan under this union is either a request or a filter
+	 * with a request.
 	 *
-	 * In such cases, this function turns the requests under UNION into xxAdd operators with the inputPlan as subplans.
+	 * Given such input plans, the function turns the requests under the union
+	 * into gpAdd operators with the first given plan as a common subplan.
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the resulting plan is indeed meant to be associated with it (for
+	 * instance, when creating a physical plan for a logical plan that is
+	 * associated with this object). Also, do not create different physical
+	 * plans with the same {@link QueryPlanningInfo} object because these
+	 * objects may later be extended with additional properties for each plan;
+	 * instead, make copies of such an object if needed.
 	 */
-	public static PhysicalPlan createPlanWithUnaryOpForUnionPlan( final PhysicalPlan inputPlan, final PhysicalPlan unionPlan ) {
+	public static PhysicalPlan createPlanWithUnaryOpForUnionPlan( final PhysicalPlan inputPlan,
+	                                                              final PhysicalPlan unionPlan,
+	                                                              final QueryPlanningInfo qpInfo ) {
 		final int numberOfSubPlansUnderUnion = unionPlan.numberOfSubPlans();
-		final PhysicalPlan[] newUnionSubPlans = new PhysicalPlan[numberOfSubPlansUnderUnion];
+		final List<PhysicalPlan> newUnionSubPlans = new ArrayList<>(numberOfSubPlansUnderUnion);
 
 		for ( int i = 0; i < numberOfSubPlansUnderUnion; i++ ) {
 			final PhysicalPlan oldSubPlan = unionPlan.getSubPlan(i);
-			final PhysicalPlan newSubPlan = createPlanWithDefaultUnaryOpIfPossible( inputPlan, oldSubPlan );
-			newUnionSubPlans[i] = newSubPlan;
+			final QueryPlanningInfo qpInfoForNewSubPlan = null;
+			final PhysicalPlan newSubPlan = createPlanWithDefaultUnaryOpIfPossible(inputPlan, oldSubPlan, qpInfoForNewSubPlan);
+			newUnionSubPlans.add(newSubPlan);
 		}
 
-		return createPlan( LogicalOpMultiwayUnion.getInstance(), newUnionSubPlans );
+		return createPlan( LogicalOpMultiwayUnion.getInstance(), qpInfo, newUnionSubPlans );
 	}
 
 	/**
-	 * If the nextPlan is in the form of a request, filter with request, or union with requests,
-	 * this function turns the requests into xxAdd operators with the inputPlan as subplans.
+	 * If the second of the two given plans is either a request, a filter
+	 * with request, or a union over requests, then this function turns the
+	 * request(s) into gpAdd operators with the first given plan as subplan.
 	 *
-	 * Otherwise, it constructs a plan with a binary join between inputPlan and nextPlan (using the default physical operator)
+	 * Otherwise, the function returns a plan with a binary join over the two
+	 * given plans (using the default physical operator).
+	 * <p>
+	 * The qpInfo argument may be <code>null</code>. Provide an actual
+	 * {@link QueryPlanningInfo} object only if this object already exists
+	 * and the resulting plan is indeed meant to be associated with it (for
+	 * instance, when creating a physical plan for a logical plan that is
+	 * associated with this object). Also, do not create different physical
+	 * plans with the same {@link QueryPlanningInfo} object because these
+	 * objects may later be extended with additional properties for each plan;
+	 * instead, make copies of such an object if needed.
 	 **/
-	public static PhysicalPlan createPlanWithDefaultUnaryOpIfPossible( final PhysicalPlan inputPlan, final PhysicalPlan nextPlan ) {
-		final PhysicalOperator oldSubPlanRootOp = nextPlan.getRootOperator();
-		if ( oldSubPlanRootOp instanceof PhysicalOpRequest ) {
-			final PhysicalOpRequest<?,?> reqOp = (PhysicalOpRequest<?,?>) oldSubPlanRootOp;
-			final UnaryLogicalOp addOp = LogicalOpUtils.createLogicalAddOpFromPhysicalReqOp(reqOp);
-			return PhysicalPlanFactory.createPlan( addOp, inputPlan);
-		}
-		else if ( oldSubPlanRootOp instanceof PhysicalOpFilter
-				&& nextPlan.getSubPlan(0).getRootOperator() instanceof PhysicalOpRequest ) {
-			final PhysicalOpFilter filterOp = (PhysicalOpFilter) oldSubPlanRootOp;
-			final PhysicalOpRequest<?,?> reqOp = (PhysicalOpRequest<?,?>) nextPlan.getSubPlan(0).getRootOperator();
+	public static PhysicalPlan createPlanWithDefaultUnaryOpIfPossible( final PhysicalPlan inputPlan,
+	                                                                   final PhysicalPlan nextPlan,
+	                                                                   final QueryPlanningInfo qpInfo ) {
+		final PhysicalOperator rootOp = nextPlan.getRootOperator();
 
+		if ( rootOp instanceof PhysicalOpRequest reqOp ) {
 			final UnaryLogicalOp addOp = LogicalOpUtils.createLogicalAddOpFromPhysicalReqOp(reqOp);
-			final PhysicalPlan addOpPlan = PhysicalPlanFactory.createPlan( addOp, inputPlan);
-
-			return PhysicalPlanFactory.createPlan( filterOp, addOpPlan);
+			return PhysicalPlanFactory.createPlan(addOp, qpInfo, inputPlan);
 		}
-		else if ( oldSubPlanRootOp instanceof PhysicalOpLocalToGlobal
-				&& nextPlan.getSubPlan(0).getRootOperator() instanceof PhysicalOpRequest ){
-			final PhysicalOpLocalToGlobal l2gPOP = (PhysicalOpLocalToGlobal) oldSubPlanRootOp;
+
+		final PhysicalOperator subPlanRootOp = nextPlan.getSubPlan(0).getRootOperator();
+
+		if (    rootOp instanceof PhysicalOpFilter filterOp
+		     && subPlanRootOp instanceof PhysicalOpRequest reqOp ) {
+			final UnaryLogicalOp addOp = LogicalOpUtils.createLogicalAddOpFromPhysicalReqOp(reqOp);
+			final PhysicalPlan addOpPlan = PhysicalPlanFactory.createPlan(addOp, inputPlan);
+			return PhysicalPlanFactory.createPlan(filterOp, qpInfo, addOpPlan);
+		}
+
+		if (    rootOp instanceof PhysicalOpLocalToGlobal l2gPOP
+		     && subPlanRootOp instanceof PhysicalOpRequest reqOp ) {
 			final LogicalOpLocalToGlobal l2gLOP = (LogicalOpLocalToGlobal) l2gPOP.getLogicalOperator();
 			final VocabularyMapping vm = l2gLOP.getVocabularyMapping();
 
 			final LogicalOpGlobalToLocal g2l = new LogicalOpGlobalToLocal(vm);
-			final PhysicalPlan newInputPlan = PhysicalPlanFactory.createPlan( new PhysicalOpGlobalToLocal(g2l), inputPlan );
-
-			final PhysicalOpRequest<?,?> reqOp = (PhysicalOpRequest<?,?>) nextPlan.getSubPlan(0).getRootOperator();
+			final PhysicalPlan newInputPlan = PhysicalPlanFactory.createPlan(g2l, inputPlan);
 
 			final UnaryLogicalOp addOp = LogicalOpUtils.createLogicalAddOpFromPhysicalReqOp(reqOp);
-			final PhysicalPlan addOpPlan = PhysicalPlanFactory.createPlan( addOp, newInputPlan);
+			final PhysicalPlan addOpPlan = PhysicalPlanFactory.createPlan(addOp, newInputPlan);
 
-			return PhysicalPlanFactory.createPlan( l2gPOP, addOpPlan );
+			return PhysicalPlanFactory.createPlan(l2gPOP, qpInfo, addOpPlan);
 		}
-		else if ( (oldSubPlanRootOp instanceof PhysicalOpBinaryUnion || oldSubPlanRootOp instanceof PhysicalOpMultiwayUnion)
-				&& PhysicalPlanFactory.checkUnaryOpApplicableToUnionPlan(nextPlan)){
-			
-			return PhysicalPlanFactory.createPlanWithUnaryOpForUnionPlan( inputPlan, nextPlan );
+
+		if (    (rootOp instanceof PhysicalOpBinaryUnion || rootOp instanceof PhysicalOpMultiwayUnion)
+		     && PhysicalPlanFactory.checkUnaryOpApplicableToUnionPlan(nextPlan) ) {
+			return PhysicalPlanFactory.createPlanWithUnaryOpForUnionPlan(inputPlan, nextPlan, qpInfo);
 		}
-		else
-			return createPlanWithJoin(inputPlan, nextPlan);
+
+		return createPlanWithJoin(inputPlan, nextPlan, qpInfo);
 	}
 
 	/**
-	 * Check whether all operators under the UNION operator belong to any of the following:
-	 * 	 - The operator is a request
-	 * 	 - If the operator is a filter, then under that filter there must be a request,
-	 * 	 - If the operator is a L2G operator, under the L2G operator, there must be a request or a filter operator with requests.
+	 * Returns <code>true</code> if the given plan has a union operator as
+	 * its root operator and all subplans under this union are of one of the
+	 * following forms:
+	 * i) a request,
+	 * ii) a filter over a request,
+	 * iii) an l2g operator over either a request or a filter with a request.
 	 */
 	public static boolean checkUnaryOpApplicableToUnionPlan( final PhysicalPlan unionPlan ){
 		final PhysicalOperator rootOp = unionPlan.getRootOperator();
-		if ( !(rootOp instanceof PhysicalOpBinaryUnion || rootOp instanceof PhysicalOpMultiwayUnion) ){
+		if (    ! (rootOp instanceof PhysicalOpBinaryUnion)
+		     && ! (rootOp instanceof PhysicalOpMultiwayUnion) ) {
 			return false;
 		}
 
 		for ( int i = 0; i < unionPlan.numberOfSubPlans(); i++ ) {
 			final PhysicalPlan subPlan = unionPlan.getSubPlan(i);
 			final PhysicalOperator subRootOp = subPlan.getRootOperator();
-			if ( !(subRootOp instanceof PhysicalOpRequest || subRootOp instanceof PhysicalOpFilter || subRootOp instanceof PhysicalOpLocalToGlobal ) ) {
+
+			if (    ! (subRootOp instanceof PhysicalOpRequest)
+			     && ! (subRootOp instanceof PhysicalOpFilter)
+			     && ! (subRootOp instanceof PhysicalOpLocalToGlobal) ) {
 				return false;
 			}
 
-			if ( subRootOp instanceof PhysicalOpLocalToGlobal ){
+			if ( subRootOp instanceof PhysicalOpLocalToGlobal ) {
 				final PhysicalPlan subSubPlan = subPlan.getSubPlan(0);
 				final PhysicalOperator subSubRootOp = subSubPlan.getRootOperator();
-				if ( !( subSubRootOp instanceof PhysicalOpRequest || subSubRootOp instanceof PhysicalOpFilter) ){
+
+				if (    ! (subSubRootOp instanceof PhysicalOpRequest)
+				     && ! (subSubRootOp instanceof PhysicalOpFilter) ) {
 					return false;
 				}
-				if ( subSubRootOp instanceof PhysicalOpFilter ){
-					if ( !( subSubPlan.getSubPlan(0).getRootOperator() instanceof PhysicalOpRequest) ){
+
+				if ( subSubRootOp instanceof PhysicalOpFilter ) {
+					final PhysicalOperator subSubSubRootOp = subSubPlan.getSubPlan(0).getRootOperator();
+					if ( ! (subSubSubRootOp instanceof PhysicalOpRequest) ) {
 						return false;
 					}
 				}
 			}
 
-			if ( subRootOp instanceof PhysicalOpFilter ){
-				if ( !( subPlan.getSubPlan(0).getRootOperator() instanceof PhysicalOpRequest) ){
+			if ( subRootOp instanceof PhysicalOpFilter ) {
+				final PhysicalOperator subSubRootOp = subPlan.getSubPlan(0).getRootOperator();
+
+				if ( ! (subSubRootOp instanceof PhysicalOpRequest) ) {
 					return false;
 				}
 			}
 		}
+
 		return true;
 	}